requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pip==21.2.2
setuptools==57.4.0
Werkzeug==2.0.1
wheel==0.36.2
smaz-py3==1.0.2
